requirements.txt
focal_loss_torch==0.1.0
joblib==1.1.0
matplotlib==3.2.2
numpy==1.21.5
pandas==1.4.2
pyodbc==4.0.32
scikit_learn==1.0.2
torch==1.12.1
tqdm==4.62.3
xgboost==1.7.4
